Move index date helper to general header date
